+++ app/slack/templates/messages.py
@@ -3935,7 +3935,7 @@ class HumanJobQuoteMessage(SlackMessage):
         allow_adjust: bool = True,
         download_translations_job_uuid: str | None = None,
         *,
-        show_quality_discount: bool = True,
+        show_quality_discount: bool = False,
         show_savings: bool = True,
         embed_additional_costs_in_line_price: bool = False,
         total_cost_label: str | None = None,